Change short to float for input image type

+++ Utilities/btkHistogramMatching.cxx
@@ -81,7 +81,7 @@ int main (int argc, char* argv[])
   bool threshold                         = thresholdArg.getValue();
   
   //ITK declaration
-  typedef short PixelType;
+  typedef float PixelType;
   const   unsigned int        Dimension = 3;
   typedef itk::Image< PixelType, Dimension >    ImageType;
   typedef ImageType::Pointer                    ImagePointer;
